op functions: tests added
