--- promql/ext/doc.go
@@ -1,8 +1,0 @@
-// Package ext extends PromQL with new functions
-//
-// Some of the following functions were ported on purpose of backward compatibility with OkQL (Okmeter Query Language).
-// Original functions documentation is available at: https://okmeter.io/misc/docs#query-language
-// Original functions source code is available at: https://fox.flant.com/okmeter/okmeter/-/blob/master/okmeter/core/eval_funcs.py
-//
-// Functions being registered in promql runtime by using special registration wrappers.
-package promqlext
